+++ doc/releasenotes.rst
@@ -19,6 +19,7 @@ Git master branch
 
 * The image-dependent pair-potential (IDPP) interpolation scheme for connecting states---i.e., in a saddle-point search---has been moved into the method :func:`ase.neb.idpp_interpolate`. This method is a more feature-rich version that that accessible via :meth:`ase.neb.NEB.interpolate`.
 
+* Write support has been added for the Vasp 5 XDATCAR file format.
 
 Version 3.19.0
 ==============